+++ packages/opencode/src/config/config.ts
@@ -1162,19 +1162,7 @@ export namespace Config {
           url: z.string().optional().describe("Enterprise URL"),
         })
         .optional(),
-      // kilocode_change start
-      commit_message: z
-        .object({
-          prompt: z
-            .string()
-            .optional()
-            .describe(
-              "Custom system prompt for AI commit message generation. When set, replaces the default conventional commits prompt entirely.",
-            ),
-        })
-        .optional()
-        .describe("Configuration for AI-generated commit messages"),
-      // kilocode_change end
+      commit_message: KilocodeConfig.CommitMessageSchema, // kilocode_change
       compaction: z
         .object({
           auto: z.boolean().optional().describe("Enable automatic compaction when context is full (default: true)"),
