+++ packages/opencode/src/kilocode/commit-message/git-context.ts
