+++ packages/opencode/src/kilocode/commit-message/index.ts
