+++ packages/opencode/src/kilocode/server/instance.ts
@@ -6,7 +6,7 @@ import { Hono } from "hono"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import z from "zod"
 import { TelemetryRoutes } from "../../server/routes/telemetry"
-import { CommitMessageRoutes } from "../../server/routes/commit-message"
+import { CommitMessageRoutes } from "./routes/commit-message"
 import { EnhancePromptRoutes } from "../../server/routes/enhance-prompt"
 import { KilocodeRoutes } from "../../server/routes/kilocode"
 import { PermissionKilocodeRoutes } from "../permission/routes"
