+++ packages/opencode/src/kilocode/server/routes/commit-message.ts
@@ -2,9 +2,9 @@ import { Hono } from "hono"
 import { describeRoute, resolver, validator } from "hono-openapi"
 import z from "zod"
 import { generateCommitMessage } from "../../commit-message"
-import { Config } from "../../config/config" // kilocode_change
-import { lazy } from "../../util/lazy"
-import { errors } from "../error"
+import { Config } from "../../../config/config"
+import { lazy } from "../../../util/lazy"
+import { errors } from "../../../server/error"
 
 export const CommitMessageRoutes = lazy(() =>
   new Hono().post(
@@ -38,11 +38,9 @@ export const CommitMessageRoutes = lazy(() =>
     ),
     async (c) => {
       const body = c.req.valid("json")
-      // kilocode_change start — read custom prompt from config
       const config = await Config.get()
       const prompt = config.commit_message?.prompt || undefined
       const result = await generateCommitMessage({ ...body, prompt })
-      // kilocode_change end
       return c.json({ message: result.message })
     },
   ),
